requirements.txt
MetaTrader5==5.0.45
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
colorama==0.4.6
pyfiglet==1.0.2
aiosqlite==0.19.0